phy: Specialize the sync sequence searching function
As searching for the sync training sequence can consume most of the computing
power in weak-signal conditions, it makes sense to allow the compiler to
optimize the search as much as possible.

On a random recording, this got me 2-3x speedup, but this will vary wildly
depending on the received signal characteristics.

This patch is a next iteration of the patch presented in
https://osmocom.org/issues/1897 , which unfortunately has multiple deadlocking
bugs.

Note: It seems that the uint-comparison method presented in the above link
does not help at all in this specialized case. Compilers are magic.